Refactor S3ProxyHandler so that handleXYZ methods can be overriden by subclass.
The purpose of making these methods to be overridable is that sub-class want to
leverage S3ProxyHandler to check authentication. And return a presigned url if
a special header is present, instead of doing actual work. This change makes it
possible for sub-class to customize before or after S3ProxyHandler actually
handle the object.

The reason to change pom.xml to use checkstyle 7.3 is to allow setting
'ignoredAnnotations'. Otherwise, it is hard to make handleXYZ methods to be
overridable.