Do not emit error code when stream is committed
If an error occurs when writing the body of a response, e.g.,
TimeoutException, previously S3Proxy would try to emit the AWS error
code.  Since this stream was already opened and closed, this failed
and Jetty emitted this warning:

Could not send response error 500: java.lang.IllegalStateException:
STREAM

References #303.