feat!: Ensure s3proxy is not run as root by default
BREAKING CHANGE

This PR change the default port from 80 to 8000 AND creates a new user "javauser" for execution so the process is not root by default. This will clear the way for a future contribution of a helm chart

feat: Publish helm chart

feat: Use helmchart to enforce non-root

A seperate PR improved the Dockerfile but introduced a risk to breaking users
Updated this PR so that non-root is enforced only in the helmchart by default.
This can of course be modified by the user if desire via values.yaml

Signed-off-by: Ryan Faircloth <ryan@dss-i.com>

chore: repair incorrect release.json on master

remove

chore: Correct source uri

chore: comment existing secret

fix: erros in template lost in merge

fix: correct issue causing regeneration of secret