Fix EncryptedBlobStore support for s3 compatible backends without eTag suffix
Some mostly S3 REST API compatible storage backends do not return the number of multipart upload parts as a suffix to the eTag as Amazon does and as the previous code expects. An example for this is NetApp StorageGRID S3 REST API.

The old code had a fallback to just assume one encrypted part, but this is actually wrong in the multipart upload case for these backends, thus this is replace in this commit.

Example eTag structure from real AWS S3 for multipart uploads with 2 parts:`xyzabc-2`

In this case the number of parts is also not present in the object metadata, as metadata is set on S3 API when starting the multipart upload, and not when completing it, thus the number of parts is not yet known at this earlier point in time.

This change adds a third fallback option to just read the number of parts from the final part padding, which is the only place guaranteed to always be present for encrypted blobs.

Only for these backends this adds an additional GET request to the backend, but only for the actual 64 bytes of the padding.